Add support for fedora images
* Removes apt repository nonesense
* Separates elements between the distros
* Cleans up the image build variables

partially implements blueprint rpm-integration

+++ scripts/files/bootstrap_init-mysql.sh
@@ -16,15 +16,16 @@ sudo mv $TEMPFILE /etc/sudoers.d/60_trove_guest
 sudo chown GUEST_USERNAME /home/GUEST_USERNAME
 
 # Copies all the trove code to the guest image
-sudo -u GUEST_USERNAME rsync -e'ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no' -avz --exclude='.*' HOST_USERNAME@NETWORK_GATEWAY:PATH_TROVE/ /home/GUEST_USERNAME/trove
+sudo -u GUEST_USERNAME rsync -e 'ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no' -avz --exclude='.*' HOST_USERNAME@NETWORK_GATEWAY:PATH_TROVE/ /home/GUEST_USERNAME/trove
 
 # Do an apt-get update since its SUPER slow first time
-apt-get update
+if [ -f /etc/debian_version ] ; then
+    apt-get update
 
-# Disable AppArmor so that trove guestagent can change the conf
-# TODO this should probably be done in the guest and then re-enabled install
-ln -s /etc/apparmor.d/usr.sbin.mysqld /etc/apparmor.d/disable/
-apparmor_parser -R /etc/apparmor.d/usr.sbin.mysqld
-
+    # Disable AppArmor so that trove guestagent can change the conf
+    # TODO this should probably be done in the guest and then re-enabled install
+    ln -s /etc/apparmor.d/usr.sbin.mysqld /etc/apparmor.d/disable/
+    apparmor_parser -R /etc/apparmor.d/usr.sbin.mysqld
+fi
 # Starts the trove guestagent (using the upstart script)
 service trove-guest start
